requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
openai-whisper==20231117
requests==2.32.3
pydub==0.25.1
python-multipart==0.0.9
